requirements.txt
Django==3.2.7
djangorestframework>=3.12,<3.13
django-cors-headers>=3.8,<4.0
django-phonenumber-field[phonenumberslite]
requests